# docs/swagger.yaml
definitions:
...
  dto.CourseOrder:
    properties:
      course:
        $ref: '#/definitions/dto.CourseProductSummary'
        description: 訂單課表
      create_at:
        description: 創建時間
        example: "2021-05-28 11:00:00"
        type: string
      id:
        description: 訂單id
        example: "202105201300687423"
        type: string
      order_status:
        description: 訂單狀態(1:等待付款/2:已付款/3:錯誤/4:取消)
        example: 1
        type: integer
      order_type:
        description: 訂單類型(1:課表購買/2:會員訂閱)
        example: 1
        type: integer
      quantity:
        description: 數量
        example: 1
        type: integer
      sale_item:
        $ref: '#/definitions/dto.SaleItem'
        description: 銷售項目
      update_at:
        description: 更新時間
        example: "2021-05-28 11:00:00"
        type: string
      user_id:
        description: 用戶id
        example: 10001
        type: integer
    type: object
...
  dto.CourseProductSummary:
    properties:
      category:
        description: 課表類別(1:有氧心肺訓練/2:間歇肌力訓練/3:重量訓練/4:阻力訓練/5:徒手訓練/6:其他)
        example: 3
        type: integer
      course_status:
        description: 課表狀態 (1:準備中/2:審核中/3:銷售中/4:退審/5:下架)
        example: 1
        type: integer
      cover:
        description: 課表封面
        example: d2w3e15d3awe.jpg
        type: string
      id:
        description: 課表 id
        example: 2
        type: integer
      level:
        description: 強度(1:初級/2:中級/3:中高級/4:高級)
        example: 3
        type: integer
      name:
        description: 課表名稱
        example: Henry課表
        type: string
      plan_count:
        description: 計畫總數
        example: 2
        type: integer
      review:
        $ref: '#/definitions/dto.ReviewStatisticSummary'
        description: 評分統計
      sale:
        $ref: '#/definitions/dto.SaleItem'
        description: 銷售項目
      sale_type:
        description: 銷售類型(1:免費課表/2:訂閱課表/3:付費課表)
        example: 2
        type: integer
      schedule_type:
        description: 排課類別(1:單一訓練/2:多項計畫)
        example: 2
        type: integer
      trainer:
        $ref: '#/definitions/dto.TrainerSummary'
        description: 教練簡介
      workout_count:
        description: 訓練總數
        example: 10
        type: integer
    type: object
...
  dto.IABProductAPIResponse:
    properties:
      acknowledgementState:
        type: integer
      consumptionState:
        type: integer
      developerPayload:
        type: string
      kind:
        type: string
      obfuscatedExternalAccountId:
        type: string
      obfuscatedExternalProfileId:
        type: string
      orderId:
        type: string
      productId:
        type: string
      purchaseState:
        type: integer
      purchaseTimeMillis:
        type: string
      purchaseToken:
        type: string
      purchaseType:
        type: integer
      quantity:
        type: integer
      regionCode:
        type: string
    type: object
  dto.IABSubscriptionAPIResponse:
    properties:
      acknowledgementState:
        type: integer
      autoRenewing:
        type: boolean
      autoResumeTimeMillis:
        type: string
      cancelReason:
        type: integer
      cancelSurveyResult:
        $ref: '#/definitions/dto.SubscriptionCancelSurveyResult'
      countryCode:
        type: string
      developerPayload:
        type: string
      emailAddress:
        type: string
      expiryTimeMillis:
        type: string
      externalAccountId:
        type: string
      familyName:
        type: string
      givenName:
        type: string
      introductoryPriceInfo:
        $ref: '#/definitions/dto.IntroductoryPriceInfo'
      kind:
        type: string
      linkedPurchaseToken:
        type: string
      obfuscatedExternalAccountId:
        type: string
      obfuscatedExternalProfileId:
        type: string
      orderId:
        type: string
      paymentState:
        description: 0. Payment pending 1. Payment received 2. Free trial 3. Pending deferred upgrade/downgrade
        type: integer
      priceAmountMicros:
        type: string
      priceChange:
        $ref: '#/definitions/dto.SubscriptionPriceChange'
      priceCurrencyCode:
        type: string
      profileId:
        type: string
      profileName:
        type: string
      promotionCode:
        type: string
      promotionType:
        type: integer
      purchaseType:
        type: integer
      startTimeMillis:
        type: string
      userCancellationTimeMillis:
        type: string
    type: object
  dto.IAPSubscribeAPIResponse:
    properties:
      appAppleId:
        type: integer
      bundleId:
        type: string
      data:
        items:
          $ref: '#/definitions/dto.SubscriptionGroupIdentifierItem'
        type: array
      environment:
        type: string
    type: object
  dto.IntroductoryPriceInfo:
    properties:
      introductoryPriceAmountMicros:
        type: string
      introductoryPriceCurrencyCode:
        type: string
      introductoryPriceCycles:
        type: integer
      introductoryPricePeriod:
        type: string
    type: object
  dto.LastTransactionsItem:
    properties:
      originalTransactionID:
        description: 原始購買的交易標識符
        type: string
      signedRenewalInfo:
        $ref: '#/definitions/dto.SignedRenewalInfo'
      signedTransactionInfo:
        $ref: '#/definitions/dto.SignedTransactionInfo'
      status:
        description: 訂閱狀態 (1:訂閱處於活動狀態/2:訂閱已過期/3:訂閱處於計費重試期/4:訂閱處於計費寬限期/5:訂閱被撤銷)
        type: integer
    type: object
...
  dto.Price:
    properties:
      currency:
        type: string
      priceMicros:
        type: string
    type: object
...
  dto.ReviewStatisticSummary:
    properties:
      amount:
        description: 評分筆數
        example: 450
        type: integer
      score_total:
        description: 評分累積
        example: 1000
        type: integer
    type: object
  dto.SaleItem:
    properties:
      id:
        description: 銷售id
        example: 1
        type: integer
      name:
        description: 銷售名稱
        example: 銅級課表
        type: string
      product_id:
        description: 產品ID
        example: com.fitness.xxx
        type: string
      twd:
        description: 台幣價格
        example: 330
        type: integer
      type:
        description: 銷售類型(1:免費課表/2:訂閱課表/3:付費課表)
        example: 3
        type: integer
    type: object
  dto.SignedRenewalInfo:
    properties:
      appAccountToken:
        description: 應用內購買的產品標識符
        type: string
      autoRenewProductId:
        description: 在下一個計費周期續訂的產品的產品標識符
        type: string
      autoRenewStatus:
        description: 自動續訂訂閱的續訂狀態(0:自動續訂已關閉，客戶已關閉訂閱的自動續訂，並且在當前訂閱期結束時不會續訂/1:自動續訂已開啟，訂閱在當前訂閱期結束時續訂)
        type: integer
      expirationIntent:
        description: 訂閱過期的原因(1:客戶取消了訂閱/2:發生計費錯誤；例如，客戶的付款信息不再有效/3:客戶不同意最近的價格上漲/4:該產品在續訂時無法購買)
        type: integer
      gracePeriodExpiresDate:
        description: 訂閱續訂的計費寬限期到期的時間
        type: integer
      isInBillingRetryPeriod:
        description: 一個布爾值，指示 App Store 是否正在嘗試自動續訂過期的訂閱
        type: boolean
      offerIdentifier:
        description: 包含促銷代碼或促銷優惠標識符的標識符
        type: string
      offerType:
        description: 促銷優惠的類型
        type: string
      originalTransactionId:
        description: 購買的原始交易標識符
        type: string
      priceIncreaseStatus:
        description: 指示客戶是否已批准訂閱價格上漲的狀態
        type: integer
      signedDate:
        description: App Store 對 JSON Web 簽名數據進行簽名的 UNIX 時間（以毫秒為單位）
        type: integer
    type: object
...
  dto.SubscribeOrder:
    properties:
      create_at:
        description: 創建時間
        example: "2021-05-28 11:00:00"
        type: string
      id:
        description: 訂單id
        example: "202105201300687423"
        type: string
      order_status:
        description: 訂單狀態(1:等待付款/2:已付款/3:錯誤/4:取消)
        example: 1
        type: integer
      order_type:
        description: 訂單類型(1:課表購買/2:會員訂閱)
        example: 1
        type: integer
      quantity:
        description: 數量
        example: 1
        type: integer
      subscribe_plan:
        $ref: '#/definitions/dto.SubscribePlan'
        description: 銷售項目
      update_at:
        description: 更新時間
        example: "2021-05-28 11:00:00"
        type: string
      user_id:
        description: 用戶id
        example: 10001
        type: integer
    type: object
  dto.SubscribePlan:
    properties:
      id:
        description: 銷售id
        example: 1
        type: integer
      name:
        description: 銷售名稱
        example: 銅級課表
        type: string
      period:
        description: 週期(1:一個月/2:二個月/3:三個月/6:六個月/12:一年/99:永久)
        example: 12
        type: integer
      product_id:
        description: 產品ID
        example: com.fitness.xxx
        type: string
      twd:
        description: 台幣價格
        example: 330
        type: integer
    type: object
  dto.SubscriptionCancelSurveyResult:
    properties:
      cancelSurveyReason:
        type: integer
      userInputCancelReason:
        type: string
    type: object
  dto.SubscriptionGroupIdentifierItem:
    properties:
      lastTransactions:
        items:
          $ref: '#/definitions/dto.LastTransactionsItem'
        type: array
      subscriptionGroupIdentifier:
        type: string
    type: object
  dto.SubscriptionPriceChange:
    properties:
      newPrice:
        $ref: '#/definitions/dto.Price'
      state:
        type: integer
    type: object
...
  dto.TrainerSummary:
    properties:
      avatar:
        description: 教練大頭照
        example: d2w3e15d3awe.jpg
        type: string
      nickname:
        description: 教練暱稱
        example: Henry教練
        type: string
      skill:
        description: 專長
        example: d2w3e15d3awe.jpg
        type: string
      user_id:
        description: 關聯的用戶id
        example: 10001
        type: integer
    type: object
...
  dto.Workout:
    properties:
      end_audio:
        description: 結束語音
        example: d2e15qwe42dw.mp3
        type: string
      equipment:
        description: 所需器材(1:無需任何器材/2:啞鈴/3:槓鈴/4:固定式器材/5:彈力繩/6:壺鈴/7:訓練椅/8:瑜珈墊/9:其他)
        example: 2,3,7
        type: string
      id:
        description: 訓練 id
        example: 1
        type: integer
      name:
        description: 訓練名稱
        example: 第一天胸肌訓練
        type: string
      start_audio:
        description: 前導語音
        example: e6d2131w5q.mp3
        type: string
      workout_set_count:
        description: 動作組數
        example: 1
        type: integer
    type: object
...
  model.Data:
    type: object
  model.ErrorResult:
    properties:
      code:
        description: 錯誤碼
        example: 9000
        type: integer
      data:
        $ref: '#/definitions/model.Data'
        description: 回傳資料
      msg:
        description: 錯誤訊息
        example: system error!
        type: string
    type: object
...
  model.SuccessResult:
    properties:
      code:
        description: 狀態碼
        example: 0
        type: integer
      data:
        $ref: '#/definitions/model.Data'
        description: 回傳資料
      msg:
        description: 成功訊息
        example: success!
        type: string
    type: object
...
  validator.CreateCourseOrderBody:
    properties:
      course_id:
        description: 課表 id
        example: 1
        type: integer
    required:
    - course_id
    type: object
...
  validator.CreateSubscribeOrderBody:
    properties:
      subscribe_plan_id:
        description: 訂閱方案 id
        example: 1
        type: integer
    required:
    - subscribe_plan_id
    type: object
...
  validator.RedeemCourseBody:
    properties:
      order_id:
        description: 訂單 id
        example: "202105201300687423"
        type: string
    required:
    - order_id
    type: object
...
  validator.VerifyGoogleReceiptBody:
    properties:
      order_id:
        description: 訂單 id
        example: "202105201300687423"
        type: string
      product_id:
        description: 產品 id
        example: com.fitness.xxx
        type: string
      receipt_data:
        description: 收據token
        example: MIJOlgYJKoZIhvcN...
        type: string
    required:
    - order_id
    - product_id
    - receipt_data
    type: object
  validator.VerifyReceiptBody:
    properties:
      order_id:
        description: 訂單 id
        example: "202105201300687423"
        type: string
      receipt_data:
        description: 收據token
        example: MIJOlgYJKoZIhvcN...
        type: string
    required:
    - order_id
    - receipt_data
    type: object
...
info:
  contact: {}
  description: 健身平台 api
  title: fitness api
paths:
...
  /v1/app_store/access_token:
    get:
      consumes:
      - application/json
      description: 取得 Apple Store api access token
      produces:
      - application/json
      responses:
        "200":
          description: 獲取成功!
          schema:
            $ref: '#/definitions/model.SuccessResult'
        "400":
          description: 獲取失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      summary: 取得 Apple Store api access token
      tags:
      - Payment_v1
  /v1/app_store/history/{original_transaction_id}:
    get:
      consumes:
      - application/json
      description: 獲取交易歷史資料(測試用)
      parameters:
      - description: 交易id
        in: path
        name: original_transaction_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 獲取成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.IAPSubscribeAPIResponse'
              type: object
        "400":
          description: 獲取失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      summary: 獲取交易歷史資料(測試用)
      tags:
      - Payment_v1
  /v1/app_store/subscriptions/{original_transaction_id}:
    get:
      consumes:
      - application/json
      description: 獲取訂閱資料(測試用)
      parameters:
      - description: 交易id
        in: path
        name: original_transaction_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 獲取成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.IAPSubscribeAPIResponse'
              type: object
        "400":
          description: 獲取失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      summary: 獲取訂閱資料(測試用)
      tags:
      - Payment_v1
...
  /v1/course_order:
    post:
      consumes:
      - application/json
      description: 創建課表訂單
      parameters:
      - description: 輸入參數
        in: body
        name: json_body
        required: true
        schema:
          $ref: '#/definitions/validator.CreateCourseOrderBody'
      produces:
      - application/json
      responses:
        "200":
          description: 成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.CourseOrder'
              type: object
        "400":
          description: 失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 創建課表訂單
      tags:
      - Payment_v1
...
  /v1/google_play/access_token:
    get:
      consumes:
      - application/json
      description: 取得 google play api access token(測試用)
      produces:
      - application/json
      responses:
        "200":
          description: 獲取成功!
          schema:
            $ref: '#/definitions/model.SuccessResult'
        "400":
          description: 獲取失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      summary: 取得 google play api access token(測試用)
      tags:
      - Payment_v1
  /v1/google_play/product/{product_id}:
    get:
      consumes:
      - application/json
      description: 取得 google play developer api product(測試用)
      parameters:
      - description: 產品id
        in: path
        name: product_id
        required: true
        type: string
      - description: 收據token
        in: query
        name: purchase_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 獲取成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.IABProductAPIResponse'
              type: object
        "400":
          description: 獲取失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      summary: 取得 google play developer api product(測試用)
      tags:
      - Payment_v1
  /v1/google_play/subscription/{product_id}:
    get:
      consumes:
      - application/json
      description: 取得 google play developer api subscription(測試用)
      parameters:
      - description: 產品id
        in: path
        name: product_id
        required: true
        type: string
      - description: 收據token
        in: query
        name: purchase_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 獲取成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.IABSubscriptionAPIResponse'
              type: object
        "400":
          description: 獲取失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      summary: 取得 google play developer api subscription(測試用)
      tags:
      - Payment_v1
...
  /v1/plan_product/{plan_id}/workouts:
    get:
      consumes:
      - application/json
      description: 獲取訓練列表
      parameters:
      - description: 計畫id
        in: path
        name: plan_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 獲取成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Workout'
                  type: array
              type: object
        "400":
          description: 獲取失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 獲取訓練列表
      tags:
      - Explore_v1
...
  /v1/redeem_course:
    post:
      consumes:
      - application/json
      description: 兌換免費課表
      parameters:
      - description: 輸入參數
        in: body
        name: json_body
        required: true
        schema:
          $ref: '#/definitions/validator.RedeemCourseBody'
      produces:
      - application/json
      responses:
        "200":
          description: 成功!
          schema:
            $ref: '#/definitions/model.SuccessResult'
        "400":
          description: 失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 兌換免費課表
      tags:
      - Payment_v1
...
  /v1/subscribe_order:
    post:
      consumes:
      - application/json
      description: 創建訂閱訂單
      parameters:
      - description: 輸入參數
        in: body
        name: json_body
        required: true
        schema:
          $ref: '#/definitions/validator.CreateSubscribeOrderBody'
      produces:
      - application/json
      responses:
        "200":
          description: 成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.SubscribeOrder'
              type: object
        "400":
          description: 失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 創建訂閱訂單
      tags:
      - Payment_v1
...
  /v1/verify_apple_receipt:
    post:
      consumes:
      - application/json
      description: 驗證apple收據
      parameters:
      - description: 輸入參數
        in: body
        name: json_body
        required: true
        schema:
          $ref: '#/definitions/validator.VerifyReceiptBody'
      produces:
      - application/json
      responses:
        "200":
          description: 成功!
          schema:
            $ref: '#/definitions/model.SuccessResult'
        "400":
          description: 失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 驗證apple收據
      tags:
      - Payment_v1
  /v1/verify_google_receipt:
    post:
      consumes:
      - application/json
      description: 驗證google收據
      parameters:
      - description: 輸入參數
        in: body
        name: json_body
        required: true
        schema:
          $ref: '#/definitions/validator.VerifyGoogleReceiptBody'
      produces:
      - application/json
      responses:
        "200":
          description: 成功!
          schema:
            $ref: '#/definitions/model.SuccessResult'
        "400":
          description: 失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 驗證google收據
      tags:
      - Payment_v1
...
securityDefinitions:
  fitness_token:
    in: header
    name: Token
    type: apiKey
swagger: "2.0"